ss: add queue wrapper around stack-attribute
This adds helper methods around push/pop stack methods in order to
represent a queue. An element is added to the end of the queue as if it
were a stack. However, an element is effectively removed from the top of
the queue by shifting everything towards the top and popping the last.

Change-Id: I0b364b95dc2b2f7999cc91826c16d003df3cae42
Signed-off-by: Christophe Bedard <christophe.bourque.bedard@ericsson.com>